html/semantics/embedded-content-0/the-video-element/readyState_during_canplaythrough.html

<!doctype html>
<html>
 <head>
  <title>video events - readyState property during canplaythrough</title>
  <script src="/resources/testharness.js"></script>
  <script /resources/testharnessreport.js></script>
  <script src="../../../../common/media.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#mediaevents">spec reference</a></p>
  <video id="v" autoplay controls>
  </video>
  <div id="log"></div>
  <script>
var t = async_test("video.readyState should be HAVE_ENOUGH_DATA during canplaythrough event", {timeout:30000});
var v = document.getElementById("v");
v.addEventListener("canplaythrough", function() {
  t.step(function() {
   assert_equals(v.readyState,
    v.HAVE_ENOUGH_DATA);
  });
  t.done();
  v.pause();
}, false);
v.src = getVideoURI("/media/movie_300") + "?" + new Date() + Math.random();
  </script>
 </body>
</html>
